// project-extender/src/types/models/Priority.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * An issue priority.
 */
export type Priority = {
  /**
   * The URL of the issue priority.
   */
  self?: string
  /**
   * The color used to indicate the issue priority.
   */
  statusColor?: string
  /**
   * The description of the issue priority.
   */
  description?: string
  /**
   * The URL of the icon for the issue priority.
   */
  iconUrl?: string
  /**
   * The name of the issue priority.
   */
  name?: string
  /**
   * The ID of the issue priority.
   */
  id?: string
  /**
   * Whether this priority is the default.
   */
  isDefault?: boolean
}